backend/requirements.txt
# Advanced AI/ML Sales Automation System
# Complete Backend Dependencies for Enterprise-Grade Sales AI

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dateutil==2.8.2

# Database & Storage
motor==3.3.2
pymongo==4.6.0
redis==5.0.1
sqlalchemy>=2.0.0
alembic>=1.12.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography>=41.0.0
bcrypt>=4.1.0

# Core AI & Machine Learning
openai==1.3.8
transformers>=4.35.0
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
sentence-transformers>=2.2.0
tokenizers>=0.15.0

# Machine Learning Algorithms
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
catboost>=1.2.0
joblib>=1.3.0

# Data Processing & Analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Advanced NLP & Text Analysis
spacy>=3.7.0
nltk>=3.8.0
textblob>=0.17.0
textdistance>=4.6.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0

# Voice AI & Audio Processing
speechrecognition>=3.10.0
pydub>=0.25.0
librosa>=0.10.0
webrtcvad>=2.0.10
openai-whisper>=20231117
elevenlabs>=0.2.0
azure-cognitiveservices-speech>=1.34.0

# Time Series & Predictive Analytics
statsmodels>=0.14.0
pmdarima>=2.0.4

# HTTP & API Communication
httpx==0.25.2
aiofiles==23.2.1
requests==2.31.0

# Communication Channels
fastapi-mail==1.4.1
twilio==8.11.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.10.0
mypy>=1.7.0

# Performance & Optimization
numba>=0.58.0
typing-extensions>=4.8.0
click>=8.1.0
tqdm>=4.66.0
humanize>=4.8.0

# Environment & Configuration
python-dotenv>=1.0.0

# Async Processing
asyncio-throttle>=1.0.2

# Optional: For development with advanced features
# tensorflow>=2.15.0
# keras>=2.15.0
# optuna>=3.4.0
# hyperopt>=0.2.7

# Note: Pin versions for production deployment
# Adjust versions based on your Python version compatibility
Pillow>=10.0.0
opencv-python==4.8.1.78

# Data Processing
openpyxl==3.1.2
python-csv==0.0.13

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
sentry-sdk==1.38.0